// ui/src/generated/api/model/dtoClarificationResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */

export interface DtoClarificationResponse {
  announce: boolean;
  answered: boolean;
  authorName: string;
  body: string;
  createdAt: string;
  fromJury: boolean;
  id: number;
  parentId?: number;
  problemId?: string;
  problemName?: string;
  replies: DtoClarificationResponse[];
  subject: string;
}